requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
certifi==2023.11.17
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
Django==4.2.6
django-allauth==65.0.2
django-extensions==3.2.3
django-polymorphic==3.1.0
exceptiongroup==1.2.2
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nba_api==1.6.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary>=2.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
whitenoise==6.7.0
wsproto==1.2.0